Resolve `NoMethodError` when using `Rack::Static` with rack >= 2.1.0
The `Rack::VCR::Transaction#vcr_response` method was assuming that
`@res.body` would respond to the `join` method. This worked when used
with rack < 2.1.0, but it wasn't part of the rack spec. According to the
spec, `body` only needs to respond to `each` and yield String values:
https://github.com/rack/rack/blame/026537fbe8baa9a9ac71417815b5be26fc28da4f/SPEC.rdoc#L266-L267

The assumption that it would respond to `join` was broken by
https://github.com/rack/rack/pull/1434/commits/8c62821f4a464858a6b6ca3c3966ec308d2bb53e
which was released in rack 2.1.0. In rack 2.1.0 and beyond, when we
request a file that's served by the `Rack::Static` middleware, the
`Rack::Response#body` will be a `Rack::Files::Iterator`, which does not
respond to `join`. As a result, a `NoMethodError` would be raised from
`Rack::VCR::Transaction#vcr_response`. The test added in this commit
fails with that error when executed on rack >= 2.1.0 without applying
the corresponding change in `transaction.rb`:
```
  with Rack::Static
    runs the HTTP request (FAILED - 1)

Failures:

  1) Rack::VCR with Rack::Static runs the HTTP request
     Failure/Error: @res.body.join(''),

     NoMethodError:
       undefined method `join' for #<Rack::Files::Iterator:0x00007f8a534d7ee0>
       ./lib/rack/vcr/transaction.rb:32:in `vcr_response'
       ./lib/rack/vcr/transaction.rb:10:in `capture'
       ./lib/rack/vcr.rb:33:in `run_request'
       ./lib/rack/vcr.rb:20:in `call'
```

To fix the error, we've ensured that we'll invoke `join` on an `Array`
(since that will respond to `join`) by simply changing `@res.body` to
`@res.body.to_enum.to_a`.